시간을 분단위로 변경해서 저장하기

+++ Source/Scene/Profile/Model/ExerciseRecordModel.swift
@@ -10,9 +10,9 @@ import Foundation
 import RealmSwift
 
 class ExerciseRecordModel: Object {
-    @Persisted var time: Double
+    @Persisted var time: Int
     
-    convenience init(time: Double) {
+    convenience init(time: Int) {
         self.init()
         self.time = time
     }
